move types and fix buildTransformChain
