+++ examples/itp-basic/src/App.tsx
@@ -30,23 +30,10 @@ export function App() {
     setInstance(
       idom($ball.current, {
         duration: 1000,
-        // delay: 1000,
-        //paused:true,
-        x: 100,
-        rotate: 360,
-        // y: 30,
-        // opacity: 0.2,
-        //y: [-20, 50],
-        //  scaleX: 2,
-        //      z:0,
-        left: "4rem",
-        //left: ["50%", "10px"],
-        //left: ["1rem","250px"],
-        //left: ["1rem", "50%"],
-        // top: [-100, 300],
-        // opacity: [0.1, 0.6],
+        opacity: 0.2,
+        x: 400,
+        y: [-20, 330],
         ease: Ease.inOutCubic,
-        onUpdate: (e) => {},
       })
     )
   }, [])
