+++ src/interpol/Interpol.ts
@@ -3,28 +3,8 @@ import { round } from "./helpers/round"
 import Ticker from "./Ticker"
 import debug from "@wbe/debug"
 import { clamp } from "./helpers/clamp"
+import { IInterpolConstruct, IUpdateParams } from "../common"
 const log = debug("interpol:Interpol")
-
-export interface IUpdateParams {
-  value: number
-  time: number
-  progress: number
-}
-
-export interface IInterpolConstruct {
-  from?: number | (() => number)
-  to?: number | (() => number)
-  duration?: number | (() => number)
-  ease?: (t: number) => number
-  reverseEase?: (t: number) => number
-  paused?: boolean
-  delay?: number
-  debug?: boolean
-  beforeStart?: () => void
-  onUpdate?: ({ value, time, progress }: IUpdateParams) => void
-  onComplete?: ({ value, time, progress }: IUpdateParams) => void
-  onRepeatComplete?: ({ value, time, progress }: IUpdateParams) => void
-}
 
 let ID = 0
